+++ changelog.md
@@ -1,6 +1,10 @@
 ====================================================================
 CHANGELOG
 ====================================================================
+## 1.1.1
+* Replace Team Member content from user meta data to media meta data.
+* New: Video Lightbox section.
+
 ## 1.1.0
 * Fixed `site-content` padding top issue.
 * New option: Disable sticky header on scroll